Multimodal_Coarse_Grained/requirements.txt
pytorch_lightning==1.3.2
transformers==4.6.0
Pillow==8.1.0
tqdm==4.56.0
ipdb==0.13.4
einops==0.3.0
pyarrow==2.0.0
pandas==1.1.5
timm==0.4.12
ftfy
opencv-python
torchmetrics==0.6.2
